Jump back in live TV
When jumping back in live TV and the jump time interval is larger than the
current play position then a jump is done to the end of the recording instead
of to the start of the recording.
Fixed by restoring a cast that was accidentally removed in
commit 7949b0d20b5dc1b63c88ef511ccbef18c71ec8bf.

+++ mythtv/libs/libmythtv/mythplayer.cpp
@@ -1448,7 +1448,7 @@ long long MythPlayer::CalcRWTime(long long rw) const
         return rw;
     }
 
-    auto seconds = secondsFromFloat((m_framesPlayed - rw) / m_videoFrameRate);
+    auto seconds = secondsFromFloat(((int64_t)m_framesPlayed - rw) / m_videoFrameRate);
     m_playerCtx->m_tvchain->JumpToNext(false, seconds);
 
     return -1;
